refactor: Storybook環境判定共通化とMSWハンドラー再構成
- isStorybookEnvironment関数をhelpers/utils/envに追加し、環境判定を共通化
- LobbyPageのモックデータを同ディレクトリのmocks.tsに移動し、コンポーネントの責務を明確化
- MSWハンドラーをtest-utils/msw/に再構成し、Firebase AuthとFirestoreで分割管理
- Firebase認証のMSWレスポンスにproviderUserInfo配列を追加（実際のAPIレスポンスに合わせる）
